tests(search): share spawn_one_shot HTTP test helper via llm_chat::test_http
Both query_expansion and contextual_chunker had an identical ~40-line
spawn_one_shot_http helper in their #[cfg(test)] modules — a bounded
localhost server that accepts one connection, reads headers +
Content-Length body, writes a canned response, and hands back the raw
request text for assertions.

Moved to llm_chat::test_http::spawn_one_shot (pub(crate), gated on
cfg(test)).  Both features' tests use the shared version.

Also removes a semantically-duplicated dev-only surface from each
feature — future LLM features get a fixture that has already been
used by two independent test suites.
